fix(testing): wait for initial timer arm
Move fake-time timer change tracking into shared timer test infrastructure and wait for self-disposing timers to be armed before advancing time. This closes the creation-event race which could miss the first one-shot tick.


Copilot-Session: f882b77b-b8ac-433e-ae07-7d16dc531cbc
